fix:cart正常。前端只要每次在Header里带上token 都不用带上userId，后端会自动识别userId注入到@RequestAttribute("userId") Integer userId

+++ backend/src/main/java/com/example/tomatomall/dao/CartItemRepository.java
@@ -10,6 +10,5 @@ import java.util.List;
 public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
     List<CartItem> findByUserId(Integer userId);
     List<CartItem> findByUserIdAndProductId(Integer userId, Integer productId);
-
-    List<CartItem> findAllByCartItemIdIn(List<Long> cartItemIds);
+    List<CartItem> findAllByCartItemIdIn(List<Integer> cartItemIds);
 }
